requirements.txt
    matplotlib==3.3.3
    numpy==1.19.5
    opencv-python==4.5.1.48
    pandas==1.2.1
    Pillow==8.1.0
    pycrypto==2.6.1
    scikit-learn==0.24.1
    scipy==1.6.0